pyproject.toml
[project]
name = "ai4mi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "nibabel>=5.3.2",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "scipy-stubs==1.16.2.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
